# static/doc/api-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Event Tracker",
    "description": "This is the API documentation for the Event Tracker App project.\n",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://18.220.2.91/v1"
    }
  ],
  "paths": {
...
    "/event/search/": {
      "get": {
        "tags": ["Event"],
        "summary": "Search Events",
        "description": "Returns a list of Events relating to a search term.",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Search term"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Event"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [{}]
      }
    },
...
    "/event/filter/": {
      "get": {
        "tags": ["Event"],
        "summary": "Get Events by Filter",
        "description": "Returns a list of Events relating to the specified filter options.",
        "parameters": [
          {
            "in": "query",
            "name": "filters",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Filter"
              }
            },
            "required": true,
            "description": "List of filters to apply"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Event"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [{}]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Venue": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        }
      },
      "Event": {
        "type": "object",
        "properties": {
          "exec_id": {
            "type": "integer"
          },
          "poster": {
            "type": "string"
          },
          "venue": {
            "$ref": "#/components/schemas/Venue"
          },
          "name": {
            "type": "string"
          },
          "artist": {
            "type": "string"
          },
          "opener": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "promoter": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "offering": {
            "type": "string"
          },
          "price": {
            "type": "string"
          },
          "ticket_link": {
            "type": "string"
          },
          "is_age_restricted": {
            "type": "boolean"
          },
          "orig_link": {
            "type": "string"
          },
          "orig_thumb": {
            "type": "string"
          }
        }
      },
      "Filter": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "condition": {
            "type": "string"
          },
          "conjugation": {
            "type": "string"
          },
          "values": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}